requirements.txt
opencv-python
numpy
pyyaml
torch
torchvision
pandas
seaborn
matplotlib
tqdm
requests